requirements.txt
streamlit
pandas
pandas-ta
plotly
alpaca-py
streamlit-autorefresh